requirements.txt
requests
toml
jira
pyyaml